Math Expression

Construct objects for representing mathematical expressions. These objects allow for both numeric and symbolic evaluations of the expressions. Only the following functions can be used to construct mathematical expressions: r allowed_math_functions. There are two alternative constructors. MathExpressionFromFunc takes a numerical function and returns a MathExpression object. MathExpressionFromString takes a character vector giving the names of the arguments and a string representing the expression, and returns a MathExpression object.

Usage

MathExpressionFromFunc(math_function)

MathExpressionFromStrings(
  expression_string,
  argument_strings = character(0L),
  include_dots = FALSE
)

Arguments

math_function

Function that takes any number of named numeric arguments and returns a numeric object. The only functions that can be used in this function include r allowed_math_functions.

expression_string

String representing the expression, referring only to the names of the arguments and the names of the following functions: r allowed_math_functions.

argument_strings

Character vector of the names of the arguments of the mathematical expression.

include_dots

Does the expression expect three dots?

Fields

string

String representing the mathematical expression.

arguments

Character vector of the names of the arguments of the mathematical expression.

Methods

$numeric_function(): Function to numerically evaluate the mathematical expression.

$symbolic_function(): Function to symbolically evaluate the mathematical expression.

Method Arguments

The argument signature depends on the expression itself, and can be found in the arguments field.