String Data

Create objects for representing names and labels in a compartmental model.

Usage

StringDataFromFrame(data)

StringDataFromDotted(labels, name)

# S3 method for StringData
print(x, ...)

Arguments

data

Data frame with names given by column names and labels by the elements of the columns.

labels

Character vector with (dot-separated) partition labels.

name

Character scalar with (dot-separated) partition name.

x

StringData object

...

Not used but present for S3 method consistency.

Methods (by generic)

print(StringData): Print out a StringData object.

Functions

StringDataFromFrame(): Construct object from a data frame without any dots in either the names or the values.

StringDataFromDotted(): Construct object from a character scalar with (dot-separated) partition names and a character vector with (dot-separated) partition labels.

Examples

path = system.file("starter_models", "seir_symp_vax", package = "macpan2")
model = ModelFiles(path)
vars = StringDataFromFrame(model$variables())
vars
#> String data object with the following $frame():
#>               Epi   Symp    SympStruc       Vax
#> 1               S        unstructured     unvax
#> 2               E        unstructured     unvax
#> 3               I   mild    component     unvax
#> 4               I severe    component     unvax
#> 5               I              vector     unvax
#> 6               I           effective     unvax
#> 7               I               total     unvax
#> 8               R        unstructured     unvax
#> 9  infectiousness   mild    component     unvax
#> 10 infectiousness severe    component     unvax
#> 11 infectiousness              vector     unvax
#> 12              N        unstructured     unvax
#> 13            foi        unstructured     unvax
#> 14              S        unstructured       vax
#> 15              E        unstructured       vax
#> 16              I   mild    component       vax
#> 17              I severe    component       vax
#> 18              I              vector       vax
#> 19              I           effective       vax
#> 20              I               total       vax
#> 21              R        unstructured       vax
#> 22 infectiousness   mild    component       vax
#> 23 infectiousness severe    component       vax
#> 24 infectiousness              vector       vax
#> 25              N        unstructured       vax
#> 26            foi        unstructured       vax
#> 27          alpha        unstructured          
#> 28           beta        unstructured          
#> 29          gamma   mild    component          
#> 30          gamma severe    component          
#> 31                       unstructured dose_rate
vars$dot()
#> String data object with the following $frame():
#>                   Epi.Symp.SympStruc.Vax
#> 1                  S..unstructured.unvax
#> 2                  E..unstructured.unvax
#> 3                 I.mild.component.unvax
#> 4               I.severe.component.unvax
#> 5                        I..vector.unvax
#> 6                     I..effective.unvax
#> 7                         I..total.unvax
#> 8                  R..unstructured.unvax
#> 9    infectiousness.mild.component.unvax
#> 10 infectiousness.severe.component.unvax
#> 11          infectiousness..vector.unvax
#> 12                 N..unstructured.unvax
#> 13               foi..unstructured.unvax
#> 14                   S..unstructured.vax
#> 15                   E..unstructured.vax
#> 16                  I.mild.component.vax
#> 17                I.severe.component.vax
#> 18                         I..vector.vax
#> 19                      I..effective.vax
#> 20                          I..total.vax
#> 21                   R..unstructured.vax
#> 22     infectiousness.mild.component.vax
#> 23   infectiousness.severe.component.vax
#> 24            infectiousness..vector.vax
#> 25                   N..unstructured.vax
#> 26                 foi..unstructured.vax
#> 27                  alpha..unstructured.
#> 28                   beta..unstructured.
#> 29                 gamma.mild.component.
#> 30               gamma.severe.component.
#> 31              ..unstructured.dose_rate