Lookup

Lookup a subset or factor index associated with a symbol, and return the index associated with that symbol.

Usage

mp_lookup(index, symbol)

Arguments

index

Index table (see mp_index).

symbol

Character string that could possibly be associated with a subset or factor of index.