Print a Table of Existing Models

Collects information from the headers of the README files in model directories and returns the results as a data frame

Usage

mp_show_models(
  dir = system.file("starter_models", package = "macpan2"),
  show_missing = FALSE,
  for_markdown = FALSE
)

show_models(
  dir = system.file("starter_models", package = "macpan2"),
  show_missing = FALSE,
  for_markdown = FALSE
)

mp_list_models(dir = system.file("starter_models", package = "macpan2"))

Arguments

dir

directory to list

show_missing

(logical) include entries for models with no README information?

for_markdown

(logical) format for rendering the table with markdown-formatted links to model readme files?

Value

a data frame containing entries Directory (model directory), Title (model title), Description (short description)

Functions

show_models(): Synonym for mp_show_models, which is preferred. Present for back-compatibility.

mp_list_models(): Return a character vector containing model names, instead of a data frame with more information about each model.

Examples

mp_show_models(show_missing = TRUE)
#>                       Directory                           Title
#> 1                     awareness                Awareness Models
#> 2                           hiv                             HIV
#> 3    lotka_volterra_competition                  Lotka-Volterra
#> 4  lotka_volterra_predator_prey                  Lotka-Volterra
#> 5                   macpan_base                     Macpan Base
#> 6                          nfds         NFDS and Vaccine Design
#> 7                          seir                      Basic SEIR
#> 8                        shiver           SHIVER = SEIR + H + V
#> 9                            si                        Basic SI
#> 10                          sir                       Basic SIR
#> 11                      sir_age              Age-stratified SIR
#> 12                    sir_demog             SIR with Demography
#> 13                 sir_mosquito            Mosquito-Vector  SIR
#> 14                   sir_waning SIR with Waning Immunity (SIRS)
#> 15                           ww                Wastewater Model
#>                                                                                                        Description
#> 1                                                                     Behaviour modifications in response to death
#> 2                                                                                               A simple HIV model
#> 3                                                                             Simple two-species competition model
#> 4                                                                                       Simple predator-prey model
#> 5                                                         Re-implementation of the McMaster group's COVID-19 model
#> 6  An ecological model using population genomics to design optimal vaccines as implemented in Colijn et al. (2020)
#> 7                                                                      Simple epidemic model with an exposed class
#> 8                                                       A modified SEIR model with Hospitalization and Vaccination
#> 9                                                                                     A very simple epidemic model
#> 10                                                                                    A very simple epidemic model
#> 11                                                                                     An age-stratified SIR model
#> 12                                                                               An SIR model with birth and death
#> 13                                                                                  SIR model for mosquito vectors
#> 14                                                                  A basic SIR model with a flow from R back to S
#> 15                                                             Macpan base with an additional wastewater component