Simulation Bounds (Experimental)

Set the simulation bounds (start time and end time) for a calibration. This is used to override the default simulation bounds taken from the observed data passed to mp_tmb_calibrator. the first date is when the first simulated time step (chosen to be before the first data point so that infectious individuals can be built up) and the second date is the last simulated time step (chosen to be after the last data point so that there can be a forecast period). the last argument gives the scale of a single time step (in this case it should always be daily).

Usage

mp_sim_bounds(sim_start, sim_end, time_scale)

Arguments

sim_start

Start time of each simulation.

sim_end

End time of each simulation.

time_scale

Qualitative description of the size of a time step. currently only "steps", "daily", and "weekly" are allowed, and but "steps" is the only recommended version as the other two are poorly tested and will throw a warning. The recommended "steps" option assumes that positive integers are used to indicate a particular point in the simulation.