Make TMB Calibrator

Construct an object that can get used to calibrate an object produced by mp_tmb_model_spec or mp_tmb_library, and possibly modified by mp_tmb_insert or mp_tmb_update.

Usage

mp_tmb_calibrator(
  spec,
  data,
  traj,
  par,
  tv = character(),
  outputs = traj,
  default = list(),
  time = NULL
)

Arguments

spec

An TMB model spec to fit to data. Such specs can be produced by mp_tmb_model_spec or mp_tmb_library, and possibly modified with mp_tmb_insert and mp_tmb_update.

data

A data frame containing trajectories to fit to and possibly time-varying parameters. The data must be of the same format as that produced by mp_trajectory.

traj

A character vector giving the names of trajectories to fit to data, or a named list of likelihood distributions specified with mp_normal, mp_log_normal, mp_neg_bin, or mp_poisson for each trajectory.

par

A character vector giving the names of parameters, either time-varying or not, to fit using trajectory match.

tv

A character vector giving the names of parameters to make time-varying according to the values in data, or a radial basis function specified with mp_rbf.

outputs

A character vector of outputs that will be generated when mp_trajectory, mp_trajectory_sd, or mp_trajectory_ensemble are called on the optimized calibrator. By default it is just the trajectories listed in traj.

default

A list of default values to use to update the defaults in the spec. By default nothing is updated. Alternatively one could use mp_tmb_update to update the spec outside of the function. Indeed such an approach is necessary if new expressions, in addition to default updates, need to be added to the spec (e.g. seasonally varying transmission).

time

Specify the start and end time of the simulated trajectories, and the time period associated with each time step. Currently the only valid choice is NULL, which takes simulation bounds from the data.

Examples

spec = mp_tmb_library("starter_models", "sir", package = "macpan2")
sim = mp_simulator(spec, 50, "infection")
data = mp_trajectory(sim)
cal = mp_tmb_calibrator(
    spec
  , data
  , traj = "infection"
  , par = "beta"
  , default = list(beta = 0.25)
)
mp_optimize(cal)
#> outer mgc:  389.7835 
#> outer mgc:  259.2344 
#> outer mgc:  40.03711 
#> outer mgc:  1.594322 
#> outer mgc:  0.002775624 
#> outer mgc:  8.444732e-09 
#> $par
#> params 
#>    0.2 
#> 
#> $objective
#> [1] 49.74796
#> 
#> $convergence
#> [1] 0
#> 
#> $iterations
#> [1] 5
#> 
#> $evaluations
#> function gradient 
#>        6        6 
#> 
#> $message
#> [1] "relative convergence (4)"
#> 
mp_tmb_coef(cal)  ## requires broom.mixed package
#> outer mgc:  8.444732e-09 
#> outer mgc:  11.74622 
#> outer mgc:  12.05664 
#> outer mgc:  31.01941 
#>     term  mat row col default  type estimate   std.error
#> 1 params beta   0   0    0.25 fixed      0.2 0.009166433

# specifying a likelihood distribution for a trajectory
cal_like = mp_tmb_calibrator(
    spec
  , data
  , traj = list(infection = mp_log_normal(sd = 1))
  , par = "beta"
  , default = list(beta = 0.25)
)
mp_optimize(cal_like)
#> outer mgc:  306.4292 
#> outer mgc:  93.0741 
#> outer mgc:  2347.071 
#> outer mgc:  2764.957 
#> outer mgc:  264.1695 
#> outer mgc:  980.5228 
#> outer mgc:  22260.99 
#> outer mgc:  8426.115 
#> outer mgc:  2366.822 
#> outer mgc:  5879.451 
#> outer mgc:  77180.86 
#> outer mgc:  29998.8 
#> outer mgc:  60570.04 
#> outer mgc:  120432.8 
#> outer mgc:  275868.3 
#> outer mgc:  636472 
#> outer mgc:  1458789 
#> outer mgc:  914951.4 
#> outer mgc:  571030.2 
#> outer mgc:  1250339 
#> outer mgc:  2735192 
#> outer mgc:  5988435 
#> outer mgc:  13101228 
#> outer mgc:  8334724 
#> outer mgc:  5299511 
#> outer mgc:  11315011 
#> outer mgc:  24156197 
#> outer mgc:  51575373 
#> outer mgc:  110108458 
#> outer mgc:  72686697 
#> outer mgc:  47980446 
#> outer mgc:  106009039 
#> outer mgc:  234215899 
#> outer mgc:  517481079 
#> outer mgc:  1143321616 
#> outer mgc:  805713777 
#> outer mgc:  567793180 
#> outer mgc:  254303209 
#> outer mgc:  458657641 
#> outer mgc:  3622236162 
#> outer mgc:  2249995370 
#> outer mgc:  920645177 
#> outer mgc:  1642677313 
#> outer mgc:  12867544976 
#> outer mgc:  13548385132 
#> outer mgc:  12058102233 
#> outer mgc:  46027377423 
#> outer mgc:  175692597634 
#> outer mgc:  92075566171 
#> outer mgc:  18928169828 
#> outer mgc:  55478312053 
#> outer mgc:  1.024199e+12 
#> outer mgc:  215778241 
#> outer mgc:  3.125478e+15 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  50 
#> Error in (function (start, objective, gradient = NULL, hessian = NULL,     ..., scale = 1, control = list(), lower = -Inf, upper = Inf) {    par <- setNames(as.double(start), names(start))    n <- length(par)    iv <- integer(78 + 3 * n)    v <- double(130 + (n * (n + 27))/2)    .Call(C_port_ivset, 2, iv, v)    if (length(control)) {        nms <- names(control)        if (!is.list(control) || is.null(nms))             stop("'control' argument must be a named list")        pos <- pmatch(nms, names(port_cpos))        if (any(nap <- is.na(pos))) {            warning(sprintf(ngettext(length(nap), "unrecognized control element named %s ignored",                 "unrecognized control elements named %s ignored"),                 paste(sQuote(nms[nap]), collapse = ", ")), domain = NA)            pos <- pos[!nap]            control <- control[!nap]        }        ivpars <- pos <= 4        vpars <- !ivpars        if (any(ivpars))             iv[port_cpos[pos[ivpars]]] <- as.integer(unlist(control[ivpars]))        if (any(vpars))             v[port_cpos[pos[vpars]]] <- as.double(unlist(control[vpars]))    }    obj <- quote(objective(.par, ...))    rho <- new.env(parent = environment())    assign(".par", par, envir = rho)    grad <- hess <- low <- upp <- NULL    if (!is.null(gradient)) {        grad <- quote(gradient(.par, ...))        if (!is.null(hessian)) {            if (is.logical(hessian))                 stop("logical 'hessian' argument not allowed.  See documentation.")            hess <- quote(hessian(.par, ...))        }    }    if (any(lower != -Inf) || any(upper != Inf)) {        low <- rep_len(as.double(lower), length(par))        upp <- rep_len(as.double(upper), length(par))    }    else low <- upp <- numeric()    .Call(C_port_nlminb, obj, grad, hess, rho, low, upp, d = rep_len(as.double(scale),         length(par)), iv, v)    iv1 <- iv[1L]    list(par = get(".par", envir = rho), objective = v[10L],         convergence = (if (iv1 %in% 3L:6L) 0L else 1L), iterations = iv[31L],         evaluations = c(`function` = iv[6L], gradient = iv[30L]),         message = if (19 <= iv1 && iv1 <= 43) {            if (any(B <- iv1 == port_cpos)) sprintf("'control' component '%s' = %g, is out of range",                 names(port_cpos)[B], v[iv1]) else sprintf("V[IV[1]] = V[%d] = %g is out of range (see PORT docu.)",                 iv1, v[iv1])        } else port_msg(iv1))})(start = c(params = 0.25, params = 0), objective = function (x = last.par[lfixed()],     ...) {    if (tracepar) {        cat("par:\n")        print(x)    }    if (!validpar(x))         return(NaN)    if (is.null(random)) {        ans <- f(x, order = 0)        if (!ADreport) {            if (is.finite(ans) && ans < value.best) {                last.par.best <<- x                value.best <<- ans            }        }    }    else {        ans <- try({            if (MCcontrol$doMC) {                ff(x, order = 0)                MC(last.par, n = MCcontrol$n, seed = MCcontrol$seed,                   order = 0)            }            else ff(x, order = 0)        }, silent = silent)        if (is.character(ans))             ans <- NaN    }    ans}, gradient = function (x = last.par[lfixed()], ...) {    if (is.null(random)) {        ans <- f(x, order = 1)    }    else {        ans <- try({            if (MCcontrol$doMC) {                ff(x, order = 0)                MC(last.par, n = MCcontrol$n, seed = MCcontrol$seed,                   order = 1)            }            else ff(x, order = 1)        }, silent = silent)        if (is.character(ans))             ans <- rep(NaN, length(x))    }    if (tracemgc)         cat("outer mgc: ", max(abs(ans)), "\n")    ans}, hessian = function (x = last.par[lfixed()], atomic = usingAtomics()) {    if (is.null(random)) {        if (!atomic)             return(f(x, order = 2))        if (is.null(ADGrad))             retape_adgrad()        return(f(x, type = "ADGrad", order = 1))    }    else {        stop("Hessian not yet implemented for models with random effects.")    }}): NA/NaN Hessian evaluation
mp_tmb_coef(cal_like)  ## requires broom.mixed package
#> outer mgc:  50 
#> outer mgc:  NaN 
#> outer mgc:  NaN 
#> outer mgc:  50 
#> outer mgc:  50 
#> outer mgc:  31.01941 
#>       term                       mat row col default  type     estimate
#> 1   params                      beta   0   0    0.25 fixed 2.000000e-01
#> 2 params.1 distr_params_sd_infection   0   0    1.00 fixed 2.220446e-16
#>   std.error
#> 1       NaN
#> 2       NaN