Print a table of contents of available models

Collects information from the headers of the README files in the model directories and returns the results as a data frame

Usage

show_models(
  dir = system.file("starter_models", package = "macpan2"),
  show_missing = FALSE,
  for_markdown = FALSE
)

Arguments

dir

directory to list

show_missing

(logical) include entries for models with no README information?

for_markdown

(logical) format for rendering the table with markdown-formatted links to model readme files?

Value

a data frame containing entries dir (model directory), title (model title), index_entry (short description)

Examples

show_models(show_missing = TRUE)
#>                             dir                    title
#> 1    lotka_volterra_competition           Lotka-Volterra
#> 2  lotka_volterra_predator_prey           Lotka-Volterra
#> 3                   macpan_base              MacPan base
#> 4                          nfds  NFDS and vaccine design
#> 5                          seir               basic SEIR
#> 6                        shiver    SHIVER = SEIR + H + V
#> 7                            si                 basic SI
#> 8                           sir                basic SIR
#> 9                     sir_demog      SIR with demography
#> 10                   sir_waning SIR with waning immunity
#> 11                           ww         Wastewater model
#>                                                                                                        index_entry
#> 1                                                                             simple two-species competition model
#> 2                                                                                       simple predator-prey model
#> 3                                                         re-implementation of the McMaster group's COVID-19 model
#> 4  An ecological model using population genomics to design optimal vaccines as implemented in Colijn et al. (2020)
#> 5                                                                     vanilla epidemic model with an exposed class
#> 6                                                                        SEIR with vaccination and hospitalization
#> 7                                                                                     a very simple epidemic model
#> 8                                                                                     a very simple epidemic model
#> 9                                                                                An SIR model with birth and death
#> 10                                                                  a basic SIR model with a flow from R back to S
#> 11                                                             Macpan base with an additional wastewater component