Flow

Specify different kinds of flows between compartments.

Usage

mp_per_capita_flow(from, to, rate, abs_rate = NULL)

mp_per_capita_inflow(from, to, rate, abs_rate = NULL)

mp_per_capita_outflow(from, rate, abs_rate = NULL)

mp_absolute_flow(from, to, rate, rate_name = NULL)

Arguments

from

String giving the name of the compartment from which the flow originates.

to

String giving the name of the compartment to which the flow is going.

rate

String giving the expression for the per-capita or absolute flow rate. Alternatively for per-capita flows, and for back compatibility, a two-sided formula with the left-hand-side giving the name of the absolute flow rate per unit time-stepand the right-hand-side giving an expression for the per-capita rate of flow from from to to.

abs_rate

String giving the name for the absolute flow rate, which will be computed as from * rate. If a formula is passed to rate (not recommended), then this abs_rate argument will be ignored.

rate_name

String giving the name for the absolute flow rate.

Details

The examples below can be mixed and matched in mp_tmb_model_spec() to produce compartmental models. Note that the symbols used below must be used in an appropriate context (e.g., if N is used for total population size, then there must be an expression like N ~ S + I + R somewhere in the model or for models with constant population size there must be a default variable, N, with a numerical value).

Functions

mp_per_capita_inflow(): Only flow into the to compartment, and do not flow out of the from compartment. The from compartment can even be a function of a set of compartments, because it will not be updated. A common construction is mp_per_capita_inflow("N", "S", "birth_rate", "birth") for adding a birth process, which involves the total population size, N, rather than a single compartment.

mp_per_capita_outflow(): Only flow out of the from compartment, without going anywhere. This is useful for removing individuals from the system (e.g., death). To keep track of the total number of dead individuals one can use mp_per_capita_flow and set to to be a compartment for these individuals (e.g., to = "D").

mp_absolute_flow(): Experimental

Examples