Simulation Bounds

Set the simulation bounds (start time and end time) for a calibration. This is used to override the default simulation bounds taken from the observed data passed to mp_tmb_calibrator.

Usage

mp_sim_bounds(sim_start, sim_end, time_scale = "steps", time_column = "time")

Arguments

sim_start

Start time of each simulation.

sim_end

End time of each simulation.

time_scale

Qualitative description of the size of a time step. The only valid values are 'steps' and 'daily'. If you would like each time step in the model to represent one day, then you should consider using 'daily'. Note that using 'daily' will require that your data represent time using a (1) Date vector, (2) character vector in YYYY-MM-DD format, or (3) integer vector that counts the number of days since some reference. Otherwise please choose 'steps', the default, and convert your time column into integer values that represent the time step that you would like in the model.

time_column

Name of the column that will identify the time at which particular values were observed.

Value

An object to be passed to the time argument of mp_tmb_calibrator.

See also